+++ lib/packages.js
@@ -34,7 +34,8 @@ var packages = {
   'ember-routing':              { trees: null,  vendorRequirements: ['router', 'route-recognizer'],
                                                requirements: ['ember-runtime', 'ember-views'] },
   'ember-application':          { trees: null,  vendorRequirements: ['dag-map'], requirements: ['ember-routing'] },
-  'ember':                      { trees: null,  requirements: ['ember-application'] }
+  'ember':                      { trees: null,  requirements: ['ember-application'] },
+  'internal-test-helpers':      { trees: null }
 };
 
 var glimmerStatus = features['ember-glimmer'];
