+++ packages/ember/tests/routing/basic_test.js
@@ -23,7 +23,7 @@ import { Transition } from 'router/transition';
 import copy from 'ember-runtime/copy';
 import { addObserver } from 'ember-metal/observer';
 import { setTemplates, set as setTemplate } from 'ember-templates/template_registry';
-import { test, asyncTest } from 'ember-glimmer/tests/utils/skip-if-glimmer';
+import { test, asyncTest } from 'internal-test-helpers/tests/skip-if-glimmer';
 
 
 var trim = jQuery.trim;
