+++ packages/ember/tests/routing/query_params_test.js
@@ -107,7 +107,7 @@ function sharedTeardown() {
   }
 }
 
-import { test } from 'ember-glimmer/tests/utils/skip-if-glimmer';
+import { test } from 'internal-test-helpers/tests/skip-if-glimmer';
 // jscs:disable
 
 QUnit.module('Routing with Query Params', {
