+++ packages/ember/tests/routing/toplevel_dom_test.js
@@ -56,7 +56,7 @@ QUnit.test('Topmost template always get an element', function() {
   equal(jQuery('#qunit-fixture > .ember-view').text(), 'hello world');
 });
 
-import { test } from 'ember-glimmer/tests/utils/skip-if-glimmer';
+import { test } from 'internal-test-helpers/tests/skip-if-glimmer';
 
 test('If topmost view has its own element, it doesn\'t get wrapped in a higher element', function() {
   App.register('view:application', EmberView.extend({
