+++ packages/ember/tests/view_instrumentation_test.js
@@ -4,7 +4,7 @@ import Application from 'ember-application/system/application';
 import { subscribe, reset } from 'ember-metal/instrumentation';
 import { compile } from 'ember-template-compiler/tests/utils/helpers';
 import { setTemplates, set as setTemplate } from 'ember-templates/template_registry';
-import { test, testModule } from 'ember-glimmer/tests/utils/skip-if-glimmer';
+import { test, testModule } from 'internal-test-helpers/tests/skip-if-glimmer';
 
 var App, $fixture;
